+++ docs/.vuepress/config.js
@@ -1,18 +1,3 @@
-// console.log(process.env);
-let externals = {};
-// if (process.env.NODE_ENV == "production") {
-//   externals = {
-//     vue: "Vue",
-//     "vue-router": "VueRouter",
-//     "element-ui": {
-//       root: "ELEMENT",
-//       commonjs: "element-ui",
-//       commonjs2: "element-ui",
-//       amd: "element-ui",
-//     },
-//   };
-// }
-
 module.exports = {
   title: "Element SaaS 扩展",
   description: "基于 Element UI 对 SaaS 业务开发常见的交互组件",
@@ -105,7 +90,6 @@ module.exports = {
     resolve: {
       extensions: [".ts", ".tsx"],
     },
-    externals,
     module: {
       rules: [
         {